Format callback URI example as code
Otherwise it automatically becomes a clickable hyperlink in the Github web view. It’s not immediately obvious to me how to change this behaviour, and while I could add a word joiner character as I did elsewhere to invisibly prevent this, that’s undesirable here where people are very likely to copy and paste this text and not expect invisible characters.